requirements.txt
appdirs==1.4.3
bottle==0.12.18
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
ecdsa==0.15
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pyasn1==0.4.8
Pygments==2.7.4
pyparsing==2.4.6
python-jose==3.1.0
pytoml==0.1.21
requests==2.24.0
retrying==1.3.3
rsa==4.6
six==1.15.0
urllib3==1.25.9
webencodings==0.5.1
